requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
